+++ CHANGELOG.md
@@ -3,5 +3,11 @@ All notable changes to the "file-icons" extension will be documented in this fil
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## 0.0.2
+* Fixed Devicons
+* Updated Octicons Font
+* Added Atom's default icons definitions
+* Colourless icon theme
+
 ## 0.0.1
 - Initial release